// facebook-api/src/facebook/models/FacebookInterfaces.ts
export interface TransformedAd {
  id?: string;
  ad_creation_time: string;
  ad_creative_bodies: string;
  ad_creative_link_captions?: string;
  ad_creative_link_titles?: string;
  ad_delivery_start_time: string;
  ad_snapshot_url?: string;
  age_country_gender_reach_breakdown: any[];
  beneficiary_payers: any[];
  eu_total_reach: number;
  languages?: string;
  page_id?: string;
  page_name: string;
  publisher_platforms: string[];
  target_ages: string[];
  target_gender: string;
  target_locations: any[];
  creative_src: string;
  full_ad_src: string;
  callToActionUrl: string;
  landing_page_src: string;
}
// ...
export enum SubfolderOptions {
  facebook_creative = 'facebook_creative/',
  facebook_landing_page = 'facebook_landing_page/',
}